--- lesson_2_2.py
@@ -1,9 +1,0 @@
-start = input("Топлива было: ")
-end = input("Топлива осталось: ")
-distance = input("Расстояние в км: ")
-
-diff = int(start) - int(end)
-
-result = (int(diff) / int(distance))*100
-
-print(f'Расход топлива на: {distance} км = {result} литров')
